+++ stock-tracker/src/store/middlwares.js
@@ -1,28 +1,8 @@
-import {
-  setChartDataAction,
-  setChartLoadingAction,
-  setChartErrorAction
-} from "../features/chart/redux/actions";
-import {
-  setCompanyOverviewAction,
-  setLoadingOverviewAction,
-  setErrorOverviewAction
-} from "../features/overview/redux/actions";
-import {
-  setLatestNewsAction,
-  setLoadingNewsAction,
-  setErrorNewsAction
-} from "../features/latestNews/redux/actions";
-import {
-  setResponseAction,
-  setLoadingKeyStatsAction,
-  setErrorKeyStatsAction
-} from "../features/keyStats/redux/actions";
-import {
-  addTopPeersAction,
-  setLoadingPeersAction,
-  setErrorPeersAction
-} from "../features/topPeers/redux/actions";
+import { setChartLoadingAction } from "../features/chart/redux/actions";
+import { setLoadingOverviewAction } from "../features/overview/redux/actions";
+import { setLoadingNewsAction } from "../features/latestNews/redux/actions";
+import { setLoadingKeyStatsAction } from "../features/keyStats/redux/actions";
+import { setLoadingPeersAction } from "../features/topPeers/redux/actions";
 import { setSuggestionsAction } from "../features/search/redux/actions";
 import {
   ADD_SYMBOL,
@@ -30,6 +10,7 @@ import {
 } from "../features/search/redux/actionTypes";
 import { SET_CHART_TIME } from "../features/chart/redux/actionTypes";
 import { INITIAL_STARTUP } from "./actionTypes";
+import { getTopSubscription } from "./subscriptions";
 
 export const searchMiddleware = socketService => store => next => action => {
   const result = next(action);
@@ -72,30 +53,6 @@ export const chartMiddleware = socketService => store => next => action => {
 export const initialStartupMiddlware = socketService => store => next => action => {
   if (action.type === INITIAL_STARTUP) {
     console.log("Application has started ");
-    const socket = socketService.create();
-    const getTopSubscription = dispatch => {
-      const unsubscribeFns = [
-        ["StockData", setResponseAction],
-        ["CompanyOverview", setCompanyOverviewAction],
-        ["LatestNews", setLatestNewsAction],
-        ["suggestions", setSuggestionsAction],
-        ["ChartData", setChartDataAction],
-        ["TopPeers", addTopPeersAction],
-        ["StockError", error => setErrorKeyStatsAction("stockData", error)],
-        [
-          "CompanyOverviewError",
-          error => setErrorOverviewAction("companiesOverview", error)
-        ],
-        ["LatestNewsError", error => setErrorNewsAction("latestNews", error)],
-        ["ChartDataError", error => setChartErrorAction("chartData", error)],
-        ["TopPeersError", error => setErrorPeersAction("topPeers", error)]
-      ].map(([event, actionCreator]) =>
-        socketService.getSocketSubscription(socket, event, payload =>
-          dispatch(actionCreator(payload))
-        )
-      );
-      return () => unsubscribeFns.forEach(fn => fn());
-    };
     getTopSubscription(store.dispatch);
   }
   const result = next(action);
